refactor: remove isomorphic-web-socket dependency
