+++ src/common/isomorphic/colored-logger.browser.js
