+++ src/common/isomorphic/colored-logger.node.js
