+++ src/common/isomorphic/web-socket.browser.js
@@ -1,0 +1,1 @@
+export default globalThis.WebSocket;
