+++ src/common/isomorphic/web-socket.node.js
@@ -1,0 +1,3 @@
+import { default as WebSocket } from 'ws';
+
+export default WebSocket;
